+++ src/lib/helpers.py
@@ -25,18 +25,6 @@ def calculate_absolute_coords(bounding_box, frame_res):
     )
 
 
-def draw_bounding_box_relative(frame, bounding_box, frame_res):
-    topleft = (
-        math.floor(bounding_box[0] * frame_res[0]),
-        math.floor(bounding_box[1] * frame_res[1]),
-    )
-    bottomright = (
-        math.floor(bounding_box[2] * frame_res[0]),
-        math.floor(bounding_box[3] * frame_res[1]),
-    )
-    return cv2.rectangle(frame, topleft, bottomright, (255, 0, 0), 3)
-
-
 def scale_bounding_box(image_size, bounding_box, target_size):
     """Scales a bounding box to target image size"""
     x1p = bounding_box[0] / image_size[0]
@@ -51,6 +39,47 @@ def scale_bounding_box(image_size, bounding_box, target_size):
     )
 
 
+def draw_bounding_box_relative(frame, bounding_box, frame_res):
+    topleft = (
+        math.floor(bounding_box[0] * frame_res[0]),
+        math.floor(bounding_box[1] * frame_res[1]),
+    )
+    bottomright = (
+        math.floor(bounding_box[2] * frame_res[0]),
+        math.floor(bounding_box[3] * frame_res[1]),
+    )
+    return cv2.rectangle(frame, topleft, bottomright, (255, 0, 0), 3)
+
+
+def draw_object(frame, obj, camera_resolution):
+    """ Draws a single pbject on supplied frame """
+    frame = draw_bounding_box_relative(
+        frame,
+        (
+            obj["relative_x1"],
+            obj["relative_y1"],
+            obj["relative_x2"],
+            obj["relative_y2"],
+        ),
+        camera_resolution,
+    )
+
+
+def draw_objects(frame, objects, camera_resolution):
+    """ Draws objects on supplied frame """
+    for obj in objects:
+        draw_object(frame, obj, camera_resolution)
+
+
+def draw_zones(frame, zones):
+    for zone in zones:
+        if zone.objects_in_zone:
+            color = (0, 255, 0)
+        else:
+            color = (0, 0, 255)
+        cv2.polylines(frame, [zone.coordinates], True, color, 3)
+
+
 def pop_if_full(queue: Queue, item: Any):
     """If queue is full, pop oldest item and put the new item"""
     try:
@@ -63,3 +92,35 @@ def pop_if_full(queue: Queue, item: Any):
 def slugify(text: str) -> str:
     """Slugify a given text."""
     return unicode_slug.slugify(text, separator="_")
+
+
+class Filter:
+    def __init__(self, object_filter):
+        self._label = object_filter.label
+        self._confidence = object_filter.confidence
+        self._width_min = object_filter.width_min
+        self._width_max = object_filter.width_max
+        self._height_min = object_filter.height_min
+        self._height_max = object_filter.height_max
+
+    def filter_confidence(self, obj):
+        if obj["confidence"] > self._confidence:
+            return True
+        return False
+
+    def filter_width(self, obj):
+        if self._width_max > obj["width"] > self._width_min:
+            return True
+        return False
+
+    def filter_height(self, obj):
+        if self._height_max > obj["height"] > self._height_min:
+            return True
+        return False
+
+    def filter_object(self, obj):
+        return (
+            self.filter_confidence(obj)
+            and self.filter_width(obj)
+            and self.filter_height(obj)
+        )
