+++ src/lib/nvr.py
@@ -3,100 +3,181 @@ import logging
 from queue import Empty, Queue
 from threading import Event, Thread
 from typing import List
-import numpy as np
 
 import cv2
 from lib.camera import FFMPEGCamera
-from lib.helpers import draw_bounding_box_relative, calculate_absolute_coords
+from lib.helpers import draw_objects, draw_zones, Filter
 from lib.motion import MotionDetection
 from lib.recorder import FFMPEGRecorder
+from lib.zones import Zone
 
 LOGGER = logging.getLogger(__name__)
 
 
-class Filter:
-    def __init__(self, object_filter):
-        self._label = object_filter.label
-        self._confidence = object_filter.confidence
-        self._width_min = object_filter.width_min
-        self._width_max = object_filter.width_max
-        self._height_min = object_filter.height_min
-        self._height_max = object_filter.height_max
+class MQTT:
+    def __init__(self, config, mqtt_queue):
+        self._logger = logging.getLogger(__name__ + "." + config.camera.name_slug)
+        self.config = config
+        self.mqtt_queue = mqtt_queue
+        self._zones = []
 
-    def filter_confidence(self, obj):
-        if obj["confidence"] > self._confidence:
-            return True
-        return False
+    def publish_sensor(self, object_detected, detections=None):
+        if self.mqtt_queue:
+            payload = {}
+            payload["state"] = "on" if object_detected else "off"
+            payload["detections"] = detections
+            json_payload = json.dumps(payload, indent=3)
 
-    def filter_width(self, obj):
-        if self._width_max > obj["width"] > self._width_min:
-            return True
-        return False
+            self.mqtt_queue.put(
+                {"topic": self.mqtt_sensor_state_topic, "payload": json_payload}
+            )
 
-    def filter_height(self, obj):
-        if self._height_max > obj["height"] > self._height_min:
-            return True
-        return False
-
-    def filter_object(self, obj):
-        return (
-            self.filter_confidence(obj)
-            and self.filter_width(obj)
-            and self.filter_height(obj)
-        )
-
-
-class Zone:
-    def __init__(self, zone, camera_resolution, config):
-        LOGGER.debug(f"Zone: {zone}")
-        self._coordinates = zone["coordinates"]
-        self._camera_resolution = camera_resolution
-        self._objects_in_zone = []
-        self._object_filters = {}
-        zone_labels = (
-            zone["labels"] if zone["labels"] else config.object_detection.labels
-        )
-        for object_filter in zone_labels:
-            self._object_filters[object_filter.label] = Filter(object_filter)
-
-    def filter_zone(self, objects):
-        self._objects_in_zone = []
-        for obj in objects:
-            if self._object_filters.get(obj["label"]) and self._object_filters[
-                obj["label"]
-            ].filter_object(obj):
-                x1, y1, x2, y2 = calculate_absolute_coords(
-                    (
-                        obj["relative_x1"],
-                        obj["relative_y1"],
-                        obj["relative_x2"],
-                        obj["relative_y2"],
-                    ),
-                    self._camera_resolution,
+    def publish_image(self, frame, zones, resolution):
+        if self.mqtt_queue:
+            draw_zones(frame.decoded_frame_mat_rgb, zones)
+            draw_objects(
+                frame.decoded_frame_mat_rgb, frame.objects, resolution,
+            )
+            ret, jpg = cv2.imencode(".jpg", frame.decoded_frame_mat_rgb)
+            if ret:
+                self.mqtt_queue.put(
+                    {"topic": self.mqtt_camera_state_topic, "payload": jpg.tobytes()}
                 )
-                if cv2.pointPolygonTest(self.coordinates, (x2, y2), False) >= 0:
-                    self._objects_in_zone.append(obj)
+
+    def on_connect(self, client):
+        client.publish(
+            self.mqtt_switch_config_topic,
+            payload=self.mqtt_switch_config_payload,
+            retain=True,
+        )
+
+        client.publish(
+            self.mqtt_sensor_config_topic,
+            payload=self.mqtt_sensor_config_payload,
+            retain=True,
+        )
+
+        client.publish(
+            self.mqtt_camera_config_topic,
+            payload=self.mqtt_camera_config_payload,
+            retain=True,
+        )
+
+        for zone in self._zones:
+            LOGGER.debug(f"Setting up MQTT for {zone}")
+            zone.on_connect(client)
+
+        subscriptions = []
+        subscriptions.append(
+            {"topic": self.mqtt_switch_command_topic, "callback": self.on_message}
+        )
+        self.publish_sensor(False)
+
+        return subscriptions
+
+    def on_message(self, message):
+        self._logger.debug(message.payload.decode())
+        self.mqtt_queue.put(
+            {
+                "topic": self.mqtt_switch_state_topic,
+                "payload": str(message.payload.decode()),
+            }
+        )
 
     @property
-    def coordinates(self):
-        return self._coordinates
+    def mqtt_switch_config_payload(self):
+        payload = {}
+        payload["name"] = self.config.camera.mqtt_name
+        payload["command_topic"] = self.mqtt_switch_command_topic
+        payload["state_topic"] = self.mqtt_switch_state_topic
+        payload["retain"] = True
+        payload["availability_topic"] = self.config.mqtt.last_will_topic
+        payload["payload_available"] = "alive"
+        payload["payload_not_available"] = "dead"
+        return json.dumps(payload, indent=3)
 
     @property
-    def objects_in_zone(self):
-        return self._objects_in_zone
+    def mqtt_switch_command_topic(self):
+        return (
+            f"{self.config.mqtt.discovery_prefix}/switch/"
+            f"{self.config.camera.mqtt_name}/set"
+        )
+
+    @property
+    def mqtt_switch_state_topic(self):
+        return (
+            f"{self.config.mqtt.discovery_prefix}/switch/"
+            f"{self.config.camera.mqtt_name}/state"
+        )
+
+    @property
+    def mqtt_switch_config_topic(self):
+        return (
+            f"{self.config.mqtt.discovery_prefix}/switch/"
+            f"{self.config.camera.mqtt_name}/config"
+        )
+
+    @property
+    def mqtt_sensor_config_payload(self):
+        payload = {}
+        payload["name"] = self.config.camera.mqtt_name
+        payload["state_topic"] = self.mqtt_sensor_state_topic
+        payload["value_template"] = "{{ value_json.state }}"
+        payload["availability_topic"] = self.config.mqtt.last_will_topic
+        payload["payload_available"] = "alive"
+        payload["payload_not_available"] = "dead"
+        payload["json_attributes_topic"] = self.mqtt_sensor_state_topic
+        return json.dumps(payload, indent=3)
+
+    @property
+    def mqtt_sensor_state_topic(self):
+        return (
+            f"{self.config.mqtt.discovery_prefix}/sensor/"
+            f"{self.config.camera.mqtt_name}/state"
+        )
+
+    @property
+    def mqtt_sensor_config_topic(self):
+        return (
+            f"{self.config.mqtt.discovery_prefix}/sensor/"
+            f"{self.config.camera.mqtt_name}/config"
+        )
+
+    @property
+    def mqtt_camera_config_payload(self):
+        payload = {}
+        payload["name"] = self.config.camera.mqtt_name
+        payload["topic"] = self.mqtt_camera_state_topic
+        payload["availability_topic"] = self.config.mqtt.last_will_topic
+        payload["payload_available"] = "alive"
+        payload["payload_not_available"] = "dead"
+        return json.dumps(payload, indent=3)
+
+    @property
+    def mqtt_camera_state_topic(self):
+        return (
+            f"{self.config.mqtt.discovery_prefix}/camera/"
+            f"{self.config.camera.mqtt_name}/image"
+        )
+
+    @property
+    def mqtt_camera_config_topic(self):
+        return (
+            f"{self.config.mqtt.discovery_prefix}/camera/"
+            f"{self.config.camera.mqtt_name}/config"
+        )
 
 
-class FFMPEGNVR(Thread):
+class FFMPEGNVR(Thread, MQTT):
     nvr_list: List[object] = []
 
     def __init__(self, config, detector, detector_queue, mqtt_queue=None):
         Thread.__init__(self)
-        self._logger = logging.getLogger(__name__ + "." + config.camera.name_slug)
+        MQTT.__init__(self, config, mqtt_queue)
         self.nvr_list.append({config.camera.mqtt_name: self})
         self._logger.debug("Initializing NVR thread")
 
         self.config = config
-        self.mqtt_queue = mqtt_queue
         self.kill_received = False
 
         self.frame_ready = Event()
@@ -229,9 +310,6 @@ class FFMPEGNVR(Thread):
         if self.idle_frames >= (self.ffmpeg.stream_fps * self.config.recorder.timeout):
             self.publish_sensor(False)
             self.recorder.stop()
-            # TODO can prolly remove this
-            with self.object_return_queue.mutex:  # Clear any objects left in queue
-                self.object_return_queue.queue.clear()
 
             if self.config.motion_detection.trigger:
                 self.scan_for_objects.clear()
@@ -247,45 +325,6 @@ class FFMPEGNVR(Thread):
             return self.object_return_queue.get_nowait()["frame"]
         except Empty:
             return None
-
-    def draw_object(self, frame, obj):
-        """ Draws a single pbject on supplied frame """
-        frame = draw_bounding_box_relative(
-            frame,
-            (
-                obj["relative_x1"],
-                obj["relative_y1"],
-                obj["relative_x2"],
-                obj["relative_y2"],
-            ),
-            self.ffmpeg.resolution,
-        )
-        return frame
-
-    def draw_objects(self, frame, objects):
-        """ Draws objects on supplied frame """
-        for obj in objects:
-            frame = draw_bounding_box_relative(
-                frame,
-                (
-                    obj["relative_x1"],
-                    obj["relative_y1"],
-                    obj["relative_x2"],
-                    obj["relative_y2"],
-                ),
-                self.ffmpeg.resolution,
-            )
-        return frame
-
-    def draw_zones(self, frame, objects):
-        for zone in self._zones:
-            if zone.objects_in_zone:
-                color = (0, 255, 0)
-            else:
-                color = (0, 0, 255)
-            cv2.polylines(frame, [zone.coordinates], True, color, 3)
-        for obj in objects:
-            frame = self.draw_object(frame, obj)
 
     def filter_objects(self, objects):
         filtered_objects = []
@@ -321,8 +360,10 @@ class FFMPEGNVR(Thread):
         if self.object_in_view:
             self.idle_frames = 0
             if not self.recorder.is_recording:
-                thumbnail = self.draw_objects(
-                    frame.decoded_frame_umat_rgb, filtered_objects,
+                thumbnail = draw_objects(
+                    frame.decoded_frame_umat_rgb,
+                    filtered_objects,
+                    self.ffmpeg.resolution,
                 )
                 self.start_recording(thumbnail)
 
@@ -360,8 +401,8 @@ class FFMPEGNVR(Thread):
             self.object_detection(processed_frame, filtered_objects)
             self.motion()
 
-            if processed_frame and processed_frame.objects:
-                self.publish_image(processed_frame)
+            if processed_frame:
+                self.publish_image(processed_frame, self._zones, self.ffmpeg.resolution)
 
             # If we are recording and no object is detected
             if self.recorder.is_recording and self.event_over():
@@ -379,145 +420,6 @@ class FFMPEGNVR(Thread):
         self.publish_sensor(value, [])
         self._object_in_view = value
         return self._object_in_view
-
-    def on_connect(self, client):
-        client.publish(
-            self.mqtt_switch_config_topic,
-            payload=self.mqtt_switch_config_payload,
-            retain=True,
-        )
-
-        client.publish(
-            self.mqtt_sensor_config_topic,
-            payload=self.mqtt_sensor_config_payload,
-            retain=True,
-        )
-
-        client.publish(
-            self.mqtt_camera_config_topic,
-            payload=self.mqtt_camera_config_payload,
-            retain=True,
-        )
-
-        subscriptions = []
-        subscriptions.append(
-            {"topic": self.mqtt_switch_command_topic, "callback": self.on_message}
-        )
-        self.publish_sensor(False)
-
-        return subscriptions
-
-    def on_message(self, message):
-        self._logger.debug(message.payload.decode())
-        self.mqtt_queue.put(
-            {
-                "topic": self.mqtt_switch_state_topic,
-                "payload": str(message.payload.decode()),
-            }
-        )
-
-    @property
-    def mqtt_switch_config_payload(self):
-        payload = {}
-        payload["name"] = self.config.camera.mqtt_name
-        payload["command_topic"] = self.mqtt_switch_command_topic
-        payload["state_topic"] = self.mqtt_switch_state_topic
-        payload["retain"] = True
-        payload["availability_topic"] = self.config.mqtt.last_will_topic
-        payload["payload_available"] = "alive"
-        payload["payload_not_available"] = "dead"
-        return json.dumps(payload, indent=3)
-
-    @property
-    def mqtt_switch_command_topic(self):
-        return (
-            f"{self.config.mqtt.discovery_prefix}/switch/"
-            f"{self.config.camera.mqtt_name}/set"
-        )
-
-    @property
-    def mqtt_switch_state_topic(self):
-        return (
-            f"{self.config.mqtt.discovery_prefix}/switch/"
-            f"{self.config.camera.mqtt_name}/state"
-        )
-
-    @property
-    def mqtt_switch_config_topic(self):
-        return (
-            f"{self.config.mqtt.discovery_prefix}/switch/"
-            f"{self.config.camera.mqtt_name}/config"
-        )
-
-    @property
-    def mqtt_sensor_config_payload(self):
-        payload = {}
-        payload["name"] = self.config.camera.mqtt_name
-        payload["state_topic"] = self.mqtt_sensor_state_topic
-        payload["value_template"] = "{{ value_json.state }}"
-        payload["availability_topic"] = self.config.mqtt.last_will_topic
-        payload["payload_available"] = "alive"
-        payload["payload_not_available"] = "dead"
-        payload["json_attributes_topic"] = self.mqtt_sensor_state_topic
-        return json.dumps(payload, indent=3)
-
-    @property
-    def mqtt_sensor_state_topic(self):
-        return (
-            f"{self.config.mqtt.discovery_prefix}/sensor/"
-            f"{self.config.camera.mqtt_name}/state"
-        )
-
-    @property
-    def mqtt_sensor_config_topic(self):
-        return (
-            f"{self.config.mqtt.discovery_prefix}/sensor/"
-            f"{self.config.camera.mqtt_name}/config"
-        )
-
-    @property
-    def mqtt_camera_config_payload(self):
-        payload = {}
-        payload["name"] = self.config.camera.mqtt_name
-        payload["topic"] = self.mqtt_camera_state_topic
-        payload["availability_topic"] = self.config.mqtt.last_will_topic
-        payload["payload_available"] = "alive"
-        payload["payload_not_available"] = "dead"
-        return json.dumps(payload, indent=3)
-
-    @property
-    def mqtt_camera_state_topic(self):
-        return (
-            f"{self.config.mqtt.discovery_prefix}/camera/"
-            f"{self.config.camera.mqtt_name}/image"
-        )
-
-    @property
-    def mqtt_camera_config_topic(self):
-        return (
-            f"{self.config.mqtt.discovery_prefix}/camera/"
-            f"{self.config.camera.mqtt_name}/config"
-        )
-
-    def publish_sensor(self, object_detected, detections=None):
-        if self.mqtt_queue:
-            payload = {}
-            payload["state"] = "on" if object_detected else "off"
-            payload["detections"] = detections
-            json_payload = json.dumps(payload, indent=3)
-
-            self.mqtt_queue.put(
-                {"topic": self.mqtt_sensor_state_topic, "payload": json_payload}
-            )
-
-    def publish_image(self, frame):
-        if self.mqtt_queue:
-            self.draw_zones(frame.decoded_frame_mat_rgb, frame.objects)
-            ret, jpg = cv2.imencode(".jpg", frame.decoded_frame_mat_rgb)
-            if ret:
-                self.mqtt_queue.put(
-                    {"topic": self.mqtt_camera_state_topic, "payload": jpg.tobytes()}
-                )
 
     def stop(self):
         self._logger.info("Stopping NVR thread")
